3_first_rag_pipeline/requirements.txt
python-dotenv
sentence-transformers
faiss-cpu
numpy
PyPDF2
openai
groq   # optional: only needed if using Groq API instead of OpenAI
